requirements.txt
pyalsaaudio
jinja2
pyspotify
